refactor css for cell
