+++ easy-ui-react/src/DataGrid/Cell.tsx
@@ -3,9 +3,9 @@ import React, { useRef } from "react";
 import { mergeProps, useFocusRing, useTableCell } from "react-aria";
 import { TableState } from "react-stately";
 import { classNames } from "../utilities/css";
-import { useDataGridRow } from "./context";
+import { useDataGridRow, useDataGridTable } from "./context";
 
-import styles from "./DataGrid.module.scss";
+import styles from "./Cell.module.scss";
 
 type CellProps<T = unknown> = {
   cell: GridNode<T>;
@@ -13,14 +13,19 @@ type CellProps<T = unknown> = {
 };
 
 export function Cell({ cell, state }: CellProps) {
+  const table = useDataGridTable();
   const row = useDataGridRow();
   const ref = useRef(null);
   const { gridCellProps } = useTableCell({ node: cell }, state, ref);
   const { isFocusVisible, focusProps } = useFocusRing();
   const className = classNames(
-    styles.cell,
-    isFocusVisible && styles.cellFocused,
-    row.isExpanded && styles.cellExpanded,
+    styles.Cell,
+    isFocusVisible && styles.focused,
+    row.isExpanded && styles.expanded,
+    table.isLeftEdgeUnderScroll && styles.leftEdgeUnderScroll,
+    table.isRightEdgeUnderScroll && styles.rightEdgeUnderScroll,
+    table.hasRowActions && styles.hasEndMatter,
+    (table.hasSelection || table.hasExpansion) && styles.hasStartMatter,
   );
   return (
     <div
