+++ easy-ui-react/src/DataGrid/ColumnHeader.module.scss
@@ -1,6 +1,3 @@
-/* stylelint-disable no-descending-specificity */
-/* stylelint-disable scss/operator-no-newline-after */
-
 @use "../styles/common" as *;
 
 .ColumnHeader {
