+++ easy-ui-react/src/DataGrid/DataGrid.module.scss
@@ -1,8 +1,4 @@
-/* stylelint-disable no-descending-specificity */
-/* stylelint-disable scss/operator-no-newline-after */
-
 @use "../styles/common" as *;
-@use "../styles/unstyled";
 
 .DataGrid {
   @include component-token(
@@ -56,16 +52,16 @@
   border: component-token("data-grid", "border-width") solid
     component-token("data-grid", "border-color");
   border-radius: design-token("shape.border_radius.md");
+  /* stylelint-disable scss/operator-no-newline-after */
   max-height: calc(
     (component-token("data-grid", "border-width") * 2) +
+      component-token("data-grid", "header-row-height") +
       (
-        component-token("data-grid", "header-row-height") +
-          (
-            component-token("data-grid", "body-row-height") *
-              component-token("data-grid", "max-rows")
-          )
+        component-token("data-grid", "body-row-height") *
+          component-token("data-grid", "max-rows")
       )
   );
+  /* stylelint-enable scss/operator-no-newline-after */
   overflow: auto;
   position: relative;
 }
@@ -94,105 +90,4 @@
     "header-text",
     design-token("color.blue.800")
   );
-}
-
-.cell {
-  display: inline-flex;
-  align-items: center;
-  padding: 0 design-token("space.2");
-
-  &:first-child {
-    padding-left: design-token("space.5");
-  }
-  &:last-child {
-    padding-right: design-token("space.5");
-  }
-
-  @include font-style("body1");
-  background: component-token("data-grid", "cell-bg");
-  color: component-token("data-grid", "cell-text");
-  height: component-token("data-grid", "body-row-height");
-  z-index: component-token("data-grid", "cell-z-index");
-}
-
-.cellFocused {
-  @include native-focus-ring;
-  outline-offset: -1px;
-}
-
-.cellExpanded {
-  margin-bottom: component-token("data-grid", "expanded-row-height");
-}
-
-.hasRowActions {
-  .cell {
-    &:last-child {
-      padding-left: design-token("space.0.5");
-      padding-right: design-token("space.2");
-      position: sticky;
-      right: component-token("data-grid", "sticky-right-offset");
-      z-index: component-token("data-grid", "cell-stuck-z-index");
-    }
-    &:nth-last-child(2) {
-      padding-right: design-token("space.1.5");
-    }
-  }
-  &.rightEdgeUnderScroll .cell:last-child::before {
-    content: "";
-    position: absolute;
-    top: 0;
-    right: 100%;
-    height: 100%;
-    width: component-token("data-grid", "sticky-shadow-size");
-    background-image: linear-gradient(
-      to left,
-      component-token("data-grid", "sticky-shadow-color"),
-      transparent
-    );
-    opacity: component-token("data-grid", "sticky-shadow-opacity");
-  }
-}
-
-.cell:nth-child(1) {
-  position: sticky;
-  left: component-token("data-grid", "sticky-left-offset");
-  z-index: component-token("data-grid", "cell-stuck-z-index");
-}
-
-.leftEdgeUnderScroll:not(.hasSelection):not(.hasExpansion) .cell:nth-child(1),
-.leftEdgeUnderScroll.hasSelection .cell:nth-child(2),
-.leftEdgeUnderScroll.hasExpansion .cell:nth-child(2) {
-  &::before {
-    content: "";
-    position: absolute;
-    top: 0;
-    left: 100%;
-    height: 100%;
-    width: component-token("data-grid", "sticky-shadow-size");
-    background-image: linear-gradient(
-      to right,
-      component-token("data-grid", "sticky-shadow-color"),
-      transparent
-    );
-    opacity: component-token("data-grid", "sticky-shadow-opacity");
-  }
-}
-
-.hasSelection,
-.hasExpansion {
-  .cell:nth-child(1) {
-    left: 0;
-    padding-left: design-token("space.2");
-    padding-right: design-token("space.1.5");
-  }
-  .cell:nth-child(2) {
-    left: 44px;
-    position: sticky;
-    z-index: component-token("data-grid", "cell-stuck-z-index");
-  }
-  .cell {
-    & + * {
-      padding-left: 0;
-    }
-  }
 }
